refactor: adapts accounts module to clean arch

+++ src/modules/accounts/useCases/updateUserAvatar/UpdateUserAvatarUseCase.ts
@@ -17,9 +17,7 @@ class UpdateUserAvatarUseCase {
   async execute({ user_id, avatar_file }: IRequest): Promise<void> {
     const user = await this.usersRepository.findById(user_id);
 
-    if (user.avatar) {
-      await deleteFile(`./tmp/avatar/${user.avatar}`);
-    }
+    if (user.avatar) await deleteFile(`./tmp/avatar/${user.avatar}`);
 
     user.avatar = avatar_file;
